Remove unneeded parameter reading code
