+++ ApsimNG/Resources/TreeViewImages/AnimalParameterSet.png
